test(stream): add basic test for stream trigger partition

+++ tests/army/stream/new_stream_basic_test.py
@@ -12,11 +12,15 @@ def main():
     parser.add_argument("window_type", choices=valid_window_type, help="Type of window to use in the SQL script.")
     parser.add_argument("--cache", action="store_true", help="Enable cache for the SQL script.")
     parser.add_argument("--history", action="store_true", help="Enable history for the SQL script.")
+    parser.add_argument("--part", action="store_true", help="Enable partition for the SQL script.")
+    parser.add_argument("--gen", action="store_true", help="Generate the SQL script without executing it.")
 
     args = parser.parse_args()
     window_type = args.window_type
     cache = args.cache
     history = args.history
+    part = args.part
+    gen = args.gen
 
     try:
         with open("basic_test.template", "r") as f:
@@ -40,7 +44,7 @@ def main():
     elif window_type == "event":
         template = template.replace(placeholder, "event_window (start with id > 0 end with id > 0)")
 
-    placeholder = "%SOURCE_TABLE%"
+    placeholder = "%CALC_SOURCE%"
     if cache:
         template = template.replace(placeholder, "%%trows")
     else:
@@ -52,9 +56,18 @@ def main():
     else:
         template = template.replace(placeholder, "")
 
+    placeholder = "%TRIGGER_SOURCE%"
+    if part:
+        template = template.replace(placeholder, "stream_trigger_st partition by tbname")
+    else:
+        template = template.replace(placeholder, "stream_trigger")
+
     try:
         with open("basic_test.sql", "w") as f:
             f.write(template)
+        if gen:
+            print("SQL script generated successfully. Exiting without execution.")
+            return
         command = ["taos", "-f", "basic_test.sql"]
         result = subprocess.run(command, capture_output=True, text=True)
         print(result.stdout)
@@ -63,7 +76,8 @@ def main():
         print(f"Error executing command: {e}")
         sys.exit(1)
     finally:
-        os.remove("basic_test.sql")
+        if not gen:
+            os.remove("basic_test.sql")
 
 if __name__ == "__main__":
     main()
